refactor: move test directory back to root
