+++ test/e2e/server.ts
@@ -4,7 +4,7 @@ import * as express from 'express';
 const app = express();
 const port = 3000;
 
-app.get('/test_integration_fixtures_async_js.js', (request, response, next) => {
+app.get('/test_integration_fixtures_async_ts.js', (request, response, next) => {
   if (
     request.query['cache-bust'] === 'true' &&
     request.query['retry-attempt'] === '5'
