+++ test/integration/fixtures/async.ts
