+++ test/integration/fixtures/index-no-import.ts
