+++ test/integration/fixtures/index.ts
