+++ tsconfig.json
@@ -1,8 +1,3 @@
 {
-  "extends": "./node_modules/gts/tsconfig-google.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "include": ["src/index.ts"]
+  "extends": "./node_modules/gts/tsconfig-google.json"
 }
